Increase reported increments to 3 in timeout test
The full-featured job opcode sequence contains a post-increment
done on kernel side that is not reported in the fence value passed
to userspace. As such, in a job that reports one too many increments
but doesn't actually timeout, userspace will see the fence threshold
be met, when it's met by the post-increment. The kernel will still
detect the job as timed out since it is waiting for the post-increment.

Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>